+++ src/main/java/com/lecture/study/biz/controller/study/StudyController.java
@@ -9,6 +9,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -27,11 +29,15 @@ public class StudyController {
     /**
      * 스터디 목록 조회
      * @param reqVO
+     * @param user
      * @return
      */
     @GetMapping("/list")
-    public ResponseEntity searchStudyInfoList(StudyReqVO reqVO) {
+    public ResponseEntity searchStudyInfoList(StudyReqVO reqVO, @AuthenticationPrincipal User user) {
         try {
+            // 로그인이 되어있는 경우, loginUserId 셋팅(좋아요 검색 관련)
+            if(user != null) reqVO.setLoginUserId(user.getUsername());
+
             PagingListVO<StudyResVO> resultList = studyService.searchStudyInfoList(reqVO);
             return ResponseEntity.ok().body(resultList);
         } catch (Exception e) {
@@ -43,11 +49,15 @@ public class StudyController {
     /**
      * 스터디 상세 정보 조회
      * @param reqVO
+     * @param user
      * @return
      */
     @GetMapping("/detail")
-    public ResponseEntity searchStudyInfo(StudyReqVO reqVO) {
+    public ResponseEntity searchStudyInfo(StudyReqVO reqVO, @AuthenticationPrincipal User user) {
         try {
+            // 로그인이 되어있는 경우, loginUserId 셋팅(좋아요 검색 관련)
+            if(user != null) reqVO.setLoginUserId(user.getUsername());
+
             StudyResVO rsult = studyService.searchStudyInfo(reqVO);
             return ResponseEntity.ok().body(rsult);
         } catch (Exception e) {
@@ -62,8 +72,11 @@ public class StudyController {
      * @return
      */
     @PostMapping("/save")
-    public ResponseEntity saveStudyInfo(@RequestBody StudySaveReqVO saveReqVO) {
+    public ResponseEntity saveStudyInfo(@RequestBody StudySaveReqVO saveReqVO, @AuthenticationPrincipal User user) {
         try {
+            if(user == null) throw new Exception("로그인이 필요한 서비스 입니다.");
+            saveReqVO.setRgsnUserId(user.getUsername());
+            saveReqVO.setAmnnUserId(user.getUsername());
             int result = studyService.saveStudyInfo(saveReqVO);
             return ResponseEntity.ok().body(result);
         } catch (Exception e) {
@@ -78,8 +91,11 @@ public class StudyController {
      * @return
      */
     @PostMapping("/delete")
-    public ResponseEntity deleteStudyInfo(@RequestBody StudySaveReqVO saveReqVO) {
+    public ResponseEntity deleteStudyInfo(@RequestBody StudySaveReqVO saveReqVO, @AuthenticationPrincipal User user) {
         try {
+            if(user == null) throw new Exception("로그인이 필요한 서비스 입니다.");
+            saveReqVO.setRgsnUserId(user.getUsername());
+            saveReqVO.setAmnnUserId(user.getUsername());
             int result = studyService.deleteStudyInfo(saveReqVO);
             return ResponseEntity.ok().body(result);
         } catch (Exception e) {
